dom/tests/mochitest/cowl/test_confinement.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display"></p>
  <pre id="test">

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  function runTestCase(testCase) {
    try {
      SpecialPowers.wrap(window["testIframe"]).eval(testCase.script);
      ok(!testCase.shouldThrow, testCase.desc);
    } catch(e) {
      ok(testCase.shouldThrow, testCase.desc);
    }

    runNextTest();
  }

  var testCaseIndex = -1;

  var testCases = [
    {
      desc: "Test 1: Try accessing localStorage after confinement",
      script: "COWL.enable(); COWL.confidentiality = new Label('http://a.com'); localStorage.setItem('secret', 'secretvalue')",
      shouldThrow: true
    },
    {
      desc: "Test 2: Should not be able to lower confidentiality label",
      script: "COWL.confidentiality = new Label();",
      shouldThrow: true
    },
    {
      desc: "Test 3: Should not be able to open auxiliary browsing context",
      script: "window.open('http://example.org');",
      shouldThrow: true
    }
  ];

  function runNextTest() {
    ++testCaseIndex;
    if (testCaseIndex == testCases.length) {
      SimpleTest.finish();
      return;
    }

    runTestCase(testCases[testCaseIndex]);
  }

  window.onload = function() {
    runNextTest();
  }

  </script>
  <iframe name="testIframe"></iframe>
</pre>
</body>
</html>
